Moving CSS to a separate folder to leave the "components" folder only for components; Removing all exports from PPC component except the component export itself. Moved all types to a separate file; Updated the demo according to new types file; Fixed type ColorProbability;
