+++ src/css/PPC.module.css
@@ -1,0 +1,53 @@
+.plot {
+    position: relative;
+    width: 100%;
+    height: 100%;
+}
+
+.gpu {
+    position: absolute;
+    left: 0;
+    top: 0;
+    z-index: 0;
+    width: 100%;
+    height: 100%;
+}
+
+.non_gpu {
+    position: absolute;
+    left: 0;
+    top: 0;
+    z-index: 1;
+    width: 100%;
+    height: 100%;
+}
+
+.toolbar {
+    pointer-events: none;
+    position: absolute;
+    right: 0;
+    top: 0;
+    z-index: 2;
+    width: auto;
+    height: 3rem;
+    display: flex;
+    justify-content: flex-end;
+    opacity: 0.0;
+}
+
+.plot:hover>.toolbar {
+    opacity: 0.4;
+    transition-timing-function: ease-in-out;
+    transition: 0.2s;
+}
+
+.plot:hover .toolbar:hover {
+    opacity: 1.0;
+    transition-timing-function: ease-in-out;
+    transition: 0.2s;
+}
+
+.toolbar_element {
+    padding: 1px;
+    pointer-events: auto;
+}
